Remove all superscript asterisks
I believe their purpose to stress that the variable belongs to a single signer.
But they're not necessary to indicate that  `pk` refers to an individual public
key because we use `aggpk` for the aggregate public key.
In case of `R`, this commit replaces the asterisks by `Rp` (for _p_artial).

Fixes #18.